source/requirements.txt
biopython>=1.79
numpy>=1.21.0
scikit-learn>=1.0.0
tqdm>=4.62.0
